Give "CollectiveVoice" code its own Django app (#3) (#600)
We create the app with startapp and add it to settings.py. We move all views and url code. So far, there are no models, so no migrations needed.
