+++ policykit/policybuilding_apps/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class PolicybuildingAppsConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'policybuilding_apps'
